+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+# 1.1.0
+* New fast-reset system that bypass the eject sequence (in config)
+* French translation added.
+
 # 1.0.0
 Initial release.